requirements.txt
Flask==2.0.3
passlib==1.7.4
psycopg2-binary==2.9.3
gunicorn
python-dotenv